new format ssl_cert.
this is very simple. ssh certificate is an public information
that is used in, for example, web server. the information will
be sent to browser/client for encryption data:
  message -> certficate/public key -> encryption -> server
    -> private key -> decrypted message
though ssl_cert can be generated by various ways, this format
is simply to store the known certificate. aka you have to set
up the key manually. there is no way to generate it randomly
at the moment.

ref.: http://tldp.org/HOWTO/SSL-Certificates-HOWTO/x64.html