slightly adjust behavior of x509 format and add spec-tests for it
So due to a bug, self signed certs were actually never a ca. So we
simply drop that default behavior and instead add an option with
which you can indicate that a cert should also become a CA.
This makes it even possible to have subCAs.

To better understand and verify (also in the future) the x509 an
extended set of tests was added.