implement (#33) - introduce profiles for easier password options.
It is possible to define profiles within the configuration file.
The idea behind profiles are to make it easy to group together
certain options for automatic password generation.

Trocla ships with a default set of profiles, which are part of
the `lib/trocla/default_config.yaml` configuration file. It is
possible to override the existing profiles within your own
configuration file, as well as adding more. Note that the
profiles part of the configuration file is merged together and
your configuration file has precedence.

The profiles part in the config is a hash where each entry
consist of a name (key) and a hash of options (value).

Profiles make it especially easy to define a preset of options
for SSL certificates as you will only need to set the
certificate specific options, while global options
such as C, O or OU can be preset within the profile.

Profiles are used by setting the profiles option to a name of
the pre-configured profiles, when passing options to the
password option. On the cli this looks like:

    trocla create foo plain 'profiles: rootpw'

It is possible to pass mutliple profiles as an array, while the
order will also reflect the precedence of the options.

Also it is possible to set a default profiles option in the
options part of the configuration file.