address concurrency corner cases & improve tests
when 2 concurrent threads or even processes are currently calculating the
same (expensive) format. This can happen in rare cases and might lead to
inconsistent state. The current solution adds some workarounds/safe_guars
to the problem, but is far from a proper solution. Given that we are not
able to provide full transactional support to password generation, our
options are limited what we can do. But we might lower the opportunity.