expand search path for sample config file to fix autopkgtest
This fixes the Debian CI tests issue as described in:

https://bugs.debian.org/830240

The problem with the Debian CI infrastructure is that it deliberately
removes the `lib` tree from the source to make sure the *installed*
libraries are tested. This, obviously, fails in the current code base
as the default config file cannot be found.

This code is unlikely to be idiomatic Ruby, and I would be happy to
see a better version. But it does seem to fix the test suite in CI
while still allowing the test suite to run correctly at build time.